requirements.txt
streamlit>=1.30
plotly>=5.0
pandas>=2.0
mysql-connector-python>=8.0
